improve SdkVersion
