+++ sentry-kotlin-multiplatform/src/commonJvmMain/kotlin/io/sentry/kotlin/multiplatform/extensions/SdkVersionExtensions.kt
@@ -1,0 +1,12 @@
+package io.sentry.kotlin.multiplatform.extensions
+
+import io.sentry.kotlin.multiplatform.protocol.SdkVersion
+import io.sentry.protocol.SdkVersion as JvmSdkVersion
+
+fun SdkVersion.toJvmSdkVersion(): JvmSdkVersion {
+    val sdk = JvmSdkVersion(this.name, this.version)
+    packages?.forEach { pkg ->
+        sdk.addPackage(pkg.name, pkg.version)
+    }
+    return sdk
+}
