+++ sentry-kotlin-multiplatform/src/commonMain/kotlin/io/sentry/kotlin/multiplatform/protocol/SdkVersion.kt
@@ -3,19 +3,22 @@ package io.sentry.kotlin.multiplatform.protocol
 import io.sentry.kotlin.multiplatform.BuildKonfig
 
 /** The SDK Interface describes the Sentry SDK and its configuration used to capture and transmit an event. */
-expect class SdkVersion(
-    name: String = BuildKonfig.SENTRY_KOTLIN_MULTIPLATFORM_SDK_NAME,
-    version: String = BuildKonfig.VERSION_NAME,
-    packages: MutableList<Package> = mutableListOf()
-) {
+data class SdkVersion(
     /** The name of the SDK. */
-    val name: String
+    val name: String = BuildKonfig.SENTRY_KOTLIN_MULTIPLATFORM_SDK_NAME,
 
     /** The version of the SDK. */
-    val version: String
+    val version: String = BuildKonfig.VERSION_NAME,
+) {
+    /** Packages used by the SDK. */
+    var packages: List<Package>? = mutableListOf()
+        private set
 
-    /** A list of packages used by the SDK. */
-    val packages: MutableList<Package>
+    fun addPackage(name: String, version: String) {
+        val mutableList = packages?.toMutableList()
+        mutableList?.add(Package(name, version))
+        packages = mutableList
+    }
 }
 
 data class Package(
